Gamemaker platform game Dark Elf
Myself and a friend created a game in gamemaker v1.4. The game is a
small platformer with typical platforming mechanics. Some custom code
execution, start and end screens for both player failure and game
completion, 3 complete levels, enemies, obstacles, and pickups. All
assets are open source and free to use, as are the sound clips.